chore: Move to a Cargo workspace
