+++ packages/ratchet-cli/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "ratchet-cli"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+clap = { version = "4.5.7", features = ["derive"] }
+ratchet = { path = "../ratchet" }
