+++ packages/ratchet/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "ratchet"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+clap = { version = "4.5.7", features = ["derive"] }
+ron = "0.8.1"
+regex = "1.10.5"
+serde = { version = "1.0.203", features = ["derive"] }
+toml = "0.8.14"
+walkdir = "2.5.0"
+seahash = "4.1.0"
