+++ packages/ratchet/src/lib.rs
@@ -3,4 +3,6 @@ mod ratchet;
 mod ratchet_file;
 mod rule;
 
+pub use crate::config::RATCHET_CONFIG;
 pub use crate::ratchet::{check, force, init, turn};
+pub use crate::ratchet_file::RATCHET_FILE;
